move builder to crate top-level
